% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmax.ml-package.R
\docType{package}
\name{pharmax.ml-package}
\alias{pharmax.ml}
\alias{pharmax.ml-package}
\title{pharmax.ml: Machine Learning for Pharmacometrics}
\description{
R-native and optional Python-backed machine learning helpers for
pharmacometric workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmax-org/pharmax}
  \item \url{https://pharmax-org.github.io/pharmax}
  \item Report bugs at \url{https://github.com/pharmax-org/pharmax/issues}
}

}
\author{
\strong{Maintainer}: Malek Okour \email{malek@zajaly.com}

}
\keyword{internal}
