% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_calibrate_conformal.R
\name{plot.pharmax_conformal_calibration}
\alias{plot.pharmax_conformal_calibration}
\title{Plot conformal calibration intervals}
\usage{
\method{plot}{pharmax_conformal_calibration}(x, ...)
}
\arguments{
\item{x}{A \code{pharmax_conformal_calibration} object.}

\item{...}{Unused.}
}
\value{
A ggplot2 object.
}
\description{
Plot conformal calibration intervals
}
