% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_covariate.R
\name{plot.pharmax_covariate}
\alias{plot.pharmax_covariate}
\title{Plot covariate screening results}
\usage{
\method{plot}{pharmax_covariate}(x, ...)
}
\arguments{
\item{x}{A \code{pharmax_covariate} object.}

\item{...}{Unused.}
}
\value{
A ggplot2 object.
}
\description{
Plot covariate screening results
}
